backend/api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = ""
authors = ["José Galinha <jbgalinha@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.70.0"
uvicorn = {extras = ["standard"], version = "^0.16.0"}
python-dotenv = "^0.19.2"
SQLAlchemy = "^1.4.28"
psycopg2-binary = "^2.9.2"
passlib = "^1.7.4"
bcrypt = "^3.2.0"
cryptography = "^36.0.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
icecream = "^2.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
